requirements.txt
astroid==2.2.5
autopep8==1.3.5
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
Django==2.1.1
dnspython==1.16.0
eventlet==0.25.0
firebase-admin==2.16.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-SocketIO==4.0.0
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
greenlet==0.4.15
grpcio==1.20.1
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
msgpack==0.6.1
Pillow==5.2.0
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.4.0
PyJWT==1.7.1
pylint==2.3.1
python-engineio==3.5.2
python-socketio==4.0.2
pytz==2019.1
refighter==0.1.0
requests==2.22.0
rsa==4.0
six==1.12.0
typed-ast==1.3.5
urllib3==1.25.2
Werkzeug==0.15.4
wrapt==1.11.1
